Conflicts:
	src/System.Reflection.Metadata/src/Resources/MetadataResources.resx
	src/System.Reflection.Metadata/src/System/Reflection/Internal/Utilities/MemoryBlock.cs
	src/System.Reflection.Metadata/src/System/Reflection/Metadata/FieldDefinition.cs
	src/System.Reflection.Metadata/src/System/Reflection/Metadata/Internal/Heaps.cs
	src/System.Reflection.Metadata/src/System/Reflection/Metadata/Internal/MetadataFlags.cs
	src/System.Reflection.Metadata/src/System/Reflection/Metadata/Internal/Tables.cs
	src/System.Reflection.Metadata/src/System/Reflection/Metadata/MetadataReader.cs

+++ src/System.IO.FileSystem/tests/File/Delete_str.cs
@@ -61,24 +61,27 @@ public class File_Delete_str
                 printerr("Error_19d4b! Incorrect exception thrown, exc==" + exc.ToString());
             }
 
-            // [] ArgumentException if argument is *.*
-            //-----------------------------------------------------------------
-            strLoc = "Loc_32453";
+            if (Interop.IsWindows)
+            {
+                // [] ArgumentException if argument is *.*
+                //-----------------------------------------------------------------
+                strLoc = "Loc_32453";
 
-            iCountTestcases++;
-            try
-            {
-                File.Delete("*.*");
-                iCountErrors++;
-                printerr("Error_4342! Expected exception not thrown");
-            }
-            catch (ArgumentException)
-            {
-            }
-            catch (Exception exc)
-            {
-                iCountErrors++;
-                printerr("Error_7777! Incorrect exception thrown, exc==" + exc.ToString());
+                iCountTestcases++;
+                try
+                {
+                    File.Delete("*.*");
+                    iCountErrors++;
+                    printerr("Error_4342! Expected exception not thrown");
+                }
+                catch (ArgumentException)
+                {
+                }
+                catch (Exception exc)
+                {
+                    iCountErrors++;
+                    printerr("Error_7777! Incorrect exception thrown, exc==" + exc.ToString());
+                }
             }
 
             // [] Exception for "."
